requirements.txt
python-telegram-bot[job-queue]==20.7
requests==2.31.0
beautifulsoup4==4.12.2
imdbpy==2023.5.1
flask==2.3.3
python-dotenv==1.0.0
yt-dlp==2023.12.30
ffmpeg-python==0.2.0
dropbox==11.36.2
